support for 64bit under mac

+++ libdivecomputer/src/serial_posix.c
@@ -531,7 +531,7 @@ serial_set_break (serial_t *device, int level)
 	if (device == NULL)
 		return -1; // EINVAL (Invalid argument)
 
-	int action = (level ? TIOCSBRK : TIOCCBRK);
+	long action = (level ? TIOCSBRK : TIOCCBRK);
 
 	if (ioctl (device->fd, action, NULL) != 0) {
 		TRACE ("ioctl");
@@ -545,7 +545,7 @@ serial_set_break (serial_t *device, int level)
 static int
 serial_set_status (int fd, int value, int level)
 {
-	int action = (level ? TIOCMBIS : TIOCMBIC);
+	long action = (level ? TIOCMBIS : TIOCMBIC);
 
 	if (ioctl (fd, action, &value) != 0) {
 		TRACE ("ioctl");
